requirements.txt
openai>=1.30.0
numpy
python-dotenv